fix: workaround unity bug where UV channels are incorrectly stripped
Unity 2022 does not seem to be able to detect that liltoon reads from
TEXCOORD1 (and later) UV channels when building asset bundles; it
therefore strips these from meshes during VRChat avatar builds.

This change adds a new fake pass to each UsePass-type liltoon shader.
While this pass is never executed (due to having LightMode=Never), it
serves to convince Unity that UV1+ is actually required. This workaround
is quite fragile; it doesn't work using the appdata structure in
lil_common_appdata, so we just open-code the whole thing without using
any includes for now.

Note that I didn't see a way to regenerate all of the built-in shaders
for a check-in; as such, only lts.shader is actually updated with this
change.

Fixes: #130