Add a mode where IDMask settings mask dissolve instead of just hiding vertices
This adds a new _IDMaskControlsDissolve flag, as well as a second set of mask
flags (_IDMaskPrior1...8).

With _IDMaskControlsDissolve true, dissolve will be enabled/disabled, and/or
its state inverted, based on the state of the IDMask flags. In particular:

* Not masked in _IDMask array, not masked in _IDMaskPrior: Not dissolved
* Masked in both _IDMask and _IDMaskPrior: Hidden by vertex shader
* Masked in only _IDMask: Dissolved
* Masked in only _IDMaskPrior: Dissolved, but the effect of dissolve on the
  main texture alpha is inverted.

The intent of this is to allow for a wipe transition where multiple parts of
multiple meshes are dissolved (to either appear or disappear) with their
dissolve positions synced. This allows for this effect to be done in units
smaller than a full mesh.